+++ src/app/driver/my-page/edit-profile/page.tsx
@@ -1,5 +1,5 @@
 import profile from '@/../public/assets/profile/img_profile_upload.svg';
-import ProfileEditDriver from '@/pages/ProfileEditDriver';
+import ProfileEditDriver from '@/_pages/ProfileEditDriver';
 
 export default function ProfileEditForDriverPage() {
   return (
