+++ src/app/normal/my-quote/edit/page.tsx
@@ -1,7 +1,7 @@
 import moveInfoData from '@/../public/data/estimationsData.json';
+import MyQuoteEditClient from '@/_pages/MyQuoteEditClient';
 import WaitingQuoteTab from '@/components/Tabs/WaitingQuoteTab';
 import EstimationInformationCard from '@/components/cards/EstimateInformationCard';
-import MyQuoteEditClient from '@/pages/MyQuoteEditClient';
 
 export default async function MyQuoteEdit() {
   return (
