+++ src/app/normal/profile-register/page.tsx
@@ -1,4 +1,4 @@
-import ProfileRegisterNormal from "@/pages/ProfileRegisterNormal";
+import ProfileRegisterNormal from '@/_pages/ProfileRegisterNormal';
 
 export default function ProfileRegisterPage() {
   return <ProfileRegisterNormal />;
