+++ src/app/normal/request-quote/page.tsx
@@ -1,4 +1,4 @@
-import RequestForQuotation from "@/pages/RequestForQuotation";
+import RequestForQuotation from '@/_pages/RequestForQuotation';
 
 export default function RequestForQuotationPage() {
   return <RequestForQuotation />;
